users_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-slugify
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
greenlet
pydantic[email]
python-multipart